fix: resolve remaining Linux clippy lints
- transmute → pointer cast for signal handler (transmutes_expressible_as_ptr_casts)
- rc as i64 → rc (unnecessary_cast, already i64 on Linux)
- Range check → RangeInclusive::contains
- Default field init → struct literal with ..Default::default()
- allow(unused_imports) for cfg(target_os = "linux") and feature-gated imports

+++ rust/plex-pg-core/src/pms_process_compat.rs
@@ -466,7 +466,7 @@ pub unsafe extern "C" fn syscall(
 
     if let Some(op) = syscall_log_label(number) {
         let err = if rc < 0 { *libc::__errno_location() } else { 0 };
-        maybe_log_event(op, rc as i64, err);
+        maybe_log_event(op, rc, err);
     }
 
     rc
